evidence: Move into internal
